+++ learn/nix/simplest-flake/liba/CHANGELOG.md
@@ -1,24 +1,30 @@
 Later
 -----
 
+- scripts to convert documentation to html&pdf
+- choose a license and apply spdx headers
+- build system enables link-time-optimization, so I don't feel bad about not providing feature flags
 - New helper macro for defining inline functions.
   Maybe like `INLINE(LIBNAME)` gets defined based on `LIBNAME_IMPLEMENTATION`.
 - syntax for tail calls, out parameters
 - prefix/suffix/infix operations on `Bytes`
 - readline (and chomp) for `Bytes` (and likelyalso helpers for `Asciiz`
 - wrap OS interface to accept `Bytes`, `Asciiz`, and/or `Text` (unicode) where relevant
+- move header imports from a.h to a.c as much as possible
+- technique: can I use x-macros to compile pattern-amtching to switches?
+
 
 v 1.0.0.2-alpha
 ---------------
 
-- Changed naming convention format. (TODO audit)
-- New naming convention for conversions. (TODO document, audit)
+- Changed naming convention format.
+- Added naming convention for conversions.
 - Added more default headers: stdlib, unistd.
-- Change basic type names to be lowercase, matching new naming conventions.
 - Changed the `Sz`, `PtrDiff` set to `USz`, `ISz`, `SSz`, with more semantic implications.
 - Added `CStr`, adding intention to `char*`.
 - New module: `Bytes` aka `Bs` for (technically mutable) slices into byte strings.
 - New module: `Asciiz`, which is `Bytes` but with the intended restriction to ascii strings that can be nul-terminated (ie contain no NUL characters).
+- Added smoke tests using nix.
 
 
 v1.0.0.1-alpha